fix(rendering): add thread-safety locks and prevent zero-id sprite cleanup
- Added std::lock_guard to all SpriteLoader getters/setters and TextureGC::clear()/addSpriteToCleanup() for thread-safety
- Changed SpriteLoader::unloaded from std::atomic<bool> to plain bool protected by mutex
- Added guard in game_sprite.cpp to skip cleanup registration when sprite id is 0
- Removed redundant g_gui.gc.updateTime() call in MapDrawer::Draw()
- Changed VisualSimilarityService::m_nextIdToIndex from uint16_t to uint32_t to match getMaxID() return type

+++ source/ui/replace_tool/visual_similarity_service.cpp
@@ -220,7 +220,7 @@ VisualSimilarityService::VisualItemData VisualSimilarityService::CalculateData(u
 
 void VisualSimilarityService::OnTimer(wxTimerEvent&) {
 	int processed = 0;
-	int maxId = g_items.getMaxID();
+	uint32_t maxId = static_cast<uint32_t>(g_items.getMaxID());
 
 	// Process fewer items per tick to prevent UI lag and handle spikes
 	while (processed < 100 && (m_nextIdToIndex <= maxId || !m_retryQueue.empty())) {
