+++ source/ui/replace_tool/visual_similarity_service.h
@@ -45,7 +45,7 @@ private:
 	std::mutex dataMutex;
 
 	wxTimer m_timer;
-	uint16_t m_nextIdToIndex;
+	uint32_t m_nextIdToIndex;
 	std::deque<uint16_t> m_retryQueue;
 	std::unordered_map<uint16_t, uint8_t> m_retryAttempts;
 	static constexpr uint8_t kMaxIndexRetries = 5;
